+++ x/incentives/keeper/genesis_test.go
@@ -187,6 +187,7 @@ func TestIncentivesInitGenesis(t *testing.T) {
 	app.IncentivesKeeper.InitGenesis(ctx, types.GenesisState{
 		Params: types.Params{
 			DistrEpochIdentifier: "week",
+			InternalUptime:       types.DefaultConcentratedUptime,
 		},
 		Gauges: expectedGauges,
 		LockableDurations: []time.Duration{